third_party/blink/web_tests/fast/lists/w3-list-styles-expected.txt

I18N Tests: list-style-type

This page tests support for the CSS list-style-type property, as defined in CSS2.0.

Note: The CSS2.0 specification does not specify in any detail the mechanism for implementing the numbering schemes defined. Expectations for behaviour are based on the version of the CSS3 Lists module current at the time this test was last updated. In most cases, however, these tests simply allow you to see whether and how a particular user agent implements these features, rather than testing against the requirements of the specification.

Check whether the numbering of the lists is what you would expect.

The start attribute is used on ol to skip to higher numbers.

lower-latin

CSS says: list-style-type: lower-latin; .

The CSS2.0 specification specifies the set of characters as "a, b, c, ... z". It doesn't specify what happens after z.

a one, should show a
b two, should show b one, expect to see a two, expect to see b three, expect to see c
-> -> a one, expect to see a
-> -> b two, expect to see b
-> -> c three, expect to see c
c three, should show c
d four, should show d
e five, should show e

z twenty six, should show z
aa twenty seven, undefined
ab twenty eight, undefined

az fifty-two, undefined
ba fifty-three, undefined
bb fifty-four, undefined

upper-latin

CSS says: list-style-type: upper-latin;

The CSS2.0 specification specifies the set of characters as "A, B, C, ... Z". It doesn't specify what happens after Z.

A one, should show A
B two, should show B one, expect to see A two, expect to see B three, expect to see C
-> -> A one, expect to see A
-> -> B two, expect to see B
-> -> C three, expect to see C
C three, should show C
D four, should show D
E five, should show E

Z twenty six, should show Z
AA twenty seven, undefined
AB twenty eight, undefined

AZ fifty-two, undefined
BA fifty-three, undefined
BB fifty-four, undefined

lower-alpha

CSS says: list-style-type: lower-alpha;

The CSS2.0 specification specifies the set of characters as "a, b, c, ... z". It doesn't specify what happens after z.

a one, should show a
b two, should show b one, expect to see a two, expect to see b three, expect to see c
-> -> a one, expect to see a
-> -> b two, expect to see b
-> -> c three, expect to see c
c three, should show c
d four, should show d
e five, should show e

z twenty six, should show z
aa twenty seven, undefined
ab twenty eight, undefined

az fifty-two, undefined
ba fifty-three, undefined
bb fifty-four, undefined

upper-alpha

CSS says: list-style-type: upper-alpha;

The CSS2.0 specification specifies the set of characters as "A, B, C, ... Z". It doesn't specify what happens after Z.

A one, should show A
B two, should show B one, expect to see A two, expect to see B three, expect to see C
-> -> A one, expect to see A
-> -> B two, expect to see B
-> -> C three, expect to see C
C three, should show C
D four, should show D
E five, should show E

Z twenty six, should show Z
AA twenty seven, undefined
AB twenty eight, undefined

AZ fifty-two, undefined
BA fifty-three, undefined
BB fifty-four, undefined

lower-greek

CSS says: list-style-type: lower-greek;

The CSS2.0 specification doesn't explicitly define the set of characters in the list, nor what happens after the list is exhausted, but does say "Lowercase classical Greek alpha, beta, gamma, ... (έ, ή, ί, ...)".

α one, may show α
β two, may show β one, may show α two, may show β three, may show γ
-> -> α one, may show α
-> -> β two, may show β
-> -> γ three, may show γ
γ three, may show γ
δ four, may show δ
ε five, may show ε

ω twenty four, may show ω
αα twenty seven, undefined
αβ twenty eight, undefined

αω forty-eight, undefined
βα forty-nine, undefined
ββ fifty, undefined

hiragana

CSS says: list-style-type: hiragana;

The CSS 2.0 specification doesn't explicitly define the set of characters in the list, nor what happens after the list is exhausted.

あ one, may show あ
い two, may show い one, may show あ two, may show い three, may show う
-> -> あ one, may show あ
-> -> い two, may show い
-> -> う three, may show う
う three, may show う
え four, may show え
お five, may show お

ん forty-eight, may show ん
ああ forty-nine, undefined
あい fifty, undefined

あん ninety-six, undefined
いあ ninety-seven, undefined
いい ninety-eight, undefined

hiragana-iroha

CSS says: list-style-type: hiragana-iroha;

The CSS 2.0 specification doesn't explicitly define the set of characters in the list, nor what happens after the list is exhausted.

い one, may show い
ろ two, may show ろ one, may show い two, may show ろ three, may show は
-> -> あ one, may show い
-> -> い two, may show ろ
-> -> う three, may show は
は three, may show は
に four, may show に
ほ five, may show ほ

す forty-seven, may show す
いい forty-eight, undefined
いろ forty-nine, undefined

いす ninety-four, undefined
ろい ninety-five, undefined
ろろ ninety-six, undefined

katakana

CSS says: list-style-type: katakana;

The CSS 2.0 specification doesn't explicitly define the set of characters in the list, nor what happens after the list is exhausted.

ア one, may show ア
イ two, may show イ one, may show ア two, may show イ three, may show ウ
-> -> ア one, may show ア
-> -> イ two, may show イ
-> -> ウ three, may show ウ
ウ three, may show ウ
エ four, may show エ
オ five, may show オ

ン forty-eight, undefined
アア forty-nine, undefined
アイ fifty, undefined

アン ninety-six, undefined
イア ninety-seven, undefined
イイ ninety-eight, undefined

katakana-iroha

CSS says: list-style-type: katakana-iroha;

The CSS 2.0 specification doesn't explicitly define the set of characters in the list, nor what happens after the list is exhausted.

イ one, may show イ
ロ two, may show ロ one, may show イ two, may show ロ three, may show ハ
-> -> あ one, may show イ
-> -> い two, may show ロ
-> -> う three, may show ハ
ハ three, may show ハ
ニ four, may show ニ
ホ five, may show ホ

ス forty-seven, may show ス
イイ forty-eight, undefined
イロ forty-nine, undefined

イス ninety-four, undefined
ロイ ninety-five, undefined
ロロ ninety-six, undefined

cjk-ideographic

CSS says: list-style-type: cjk-ideographic;

The CSS2.0 specification doesn't define what characters should be used to create each number - there are multiple alternatives. Note that the characters specified in the CSS3 Lists module as of the date of publication of this page are not those expected below.

一 one
二 two one two three
-> -> 一 one
-> -> 二 two
-> -> 三 three
三 three
四 four
五 five

二十 twenty, undefined
二十一 twenty one, undefined
二十二 twenty two, undefined

一百 a hundred, undefined
一百零一 a hundred and one, undefined
一百零二 a hundred and two, undefined

零 zero, undefined
一 one

一千 a thousand, undefined

一千萬零一千 ten million one thousand, undefined

一百萬零一百 one million one hundred, undefined

hebrew

CSS says: list-style-type: hebrew;

The CSS2.0 specification doesn't define the algorithm for the Hebrew numbering.

א one, should show א
ב two, should show ב one, should show א two, should show ב three, should show ג
-> -> א one, should show א
-> -> ב two, should show ב
-> -> ג three, should show ג
ג three, should show ג
ד four, should show ד
ה five, should show ה

צז ninety-seven, undefined, may show צז
צח ninety-eight, undefined, may show צח
צט ninety-nine, undefined, may show צט
ק hundred, undefined, may show ק
קא a hundred and one, undefined, may show קא

אפס zero, undefined, may show אפס
א one, should show א

georgian

CSS says: list-style-type: georgian;

The CSS2.0 specification doesn't define the algorithm for Georgian numbering in detail.

ა one, may show ა
ბ two, may show ბ one, may show ა two, may show ბ three, may show გ
-> -> ა one, may show ა
-> -> ბ two, may show ბ
-> -> გ three, may show გ
გ three, may show გ
დ four, may show დ
ე five, may show ე
ვ six, may show ვ
ზ seven, may show ზ
ჱ eight, may show ჱ
თ nine, may show თ
ი ten, may show ი
ია eleven, may show ია
იბ twelve, may show იბ

ჟთ ninety-nine, undefined, may show ჟთ
რ hundred, undefined, may show რ
რა a hundred and one, undefined, may show რა

armenian

CSS says: list-style-type: armenian;

The CSS2.0 specification doesn't define the algorithm for Armenian numbering in detail.
